refactor: move CommunityVoteStats and VoteStats interfaces to types dir
